Task: Use getFunctionCallback in invalidate()
