+++ lib/ceych.js
@@ -5,6 +5,7 @@ const Catbox = require('catbox').Client;
 const Memory = require('catbox-memory');
 const memoize = require('./memoize');
 const createCacheKey = require('./utils').createCacheKey;
+const getFunctionCallback = require('./utils').getFunctionCallback;
 
 function validateClientOpts(opts) {
   if (!opts) {
@@ -86,7 +87,7 @@ Ceych.prototype.disableCache = function() {
 
 Ceych.prototype.invalidate = function(fn) {
   const args = Array.from(arguments);
-  const cb = args.pop();
+  const cb = getFunctionCallback(args);
   const suffix = getWrapOpts(this.defaultTTL, args);
   const cacheKey = createCacheKey(fn, args.slice(1), suffix);
   this.cache.drop(cacheKey, cb);
