+++ lib/memoize.js
@@ -3,6 +3,7 @@
 const _ = require('lodash');
 const async = require('async');
 const createCacheKey = require('./utils').createCacheKey;
+const getFunctionCallback = require('./utils').getFunctionCallback;
 
 function getFunctionArgs(_args) {
   const args = Array.prototype.slice.call(_args);
@@ -12,16 +13,6 @@ function getFunctionArgs(_args) {
   }
 
   return args;
-}
-
-function getFunctionCallback(args) {
-  const callback = args.pop();
-
-  if (typeof callback !== 'function') {
-    throw new Error('Final argument of wrapped function must be a callback');
-  }
-
-  return callback;
 }
 
 module.exports = function (cacheClient, cacheOpts, func) {
