+++ lib/utils.js
@@ -20,6 +20,17 @@ function createCacheKey(func, args, suffix) {
   };
 }
 
+function getFunctionCallback(args) {
+  const callback = args.pop();
+
+  if (typeof callback !== 'function') {
+    throw new Error('Final argument of wrapped function must be a callback');
+  }
+
+  return callback;
+}
+
 module.exports = {
+  getFunctionCallback,
   createCacheKey
 };
